% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpi.R
\docType{package}
\name{rpi}
\alias{rpi}
\alias{rpi-package}
\title{Tools and Utilities for R in Raspberry Pi}
\description{
With the ready availability of microcomputers such as the Raspberry Pi,
interest in homegrown development of \strong{internet of things} devices embedded
with various sensors and technologies has grown allowing for the much wider
adoption of citizen science - scientific research conducted, in whole
or in part, by amateur (or non-professional) scientists. The use of R in
Raspberry Pi allows the leveraging of the language's innate data science and
statistical capabilities in these citizen science projects. This package
provides functions designed with the Raspberry Pi in mind and for use in
some of the most common IoT and citizen science projects currently deployed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ernest.guevarra.io/rpi/}
  \item \url{https://github.com/ernestguevarra/rpi}
  \item Report bugs at \url{https://github.com/ernestguevarra/rpi/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
