% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds011.R
\name{sds011_connect}
\alias{sds011_connect}
\title{Connect to the SDS011 air particulate sensor connected via USB}
\usage{
sds011_connect(
  port = "ttyUSB0",
  mode = "115200,n,8,1",
  buffering = "none",
  newline = TRUE,
  translation = "auto",
  open = TRUE
)
}
\arguments{
\item{port}{Part name. This is usually operating software dependent. For
Raspberry Pi with an SDS011 device plugged via USB, this will be \emph{ttyUSB0}.}

\item{mode}{A string concatenating the baud rate, parity, databits and
stopbits of the communication mode separated by a comma. Default to
\emph{"115200,n,8,1"}.}

\item{buffering}{Connection buffering. Default to \emph{"none"}. Other options
is \emph{"line"} to send buffer after new line character or \emph{"full"} where
write operations are buffed until a connection is flushed.}

\item{newline}{Logical. Should a transmission end with a newline or not.
Default is TRUE or 1.}

\item{translation}{End of line character to use. Could be \emph{"lf"}, \emph{"cr"},
\emph{"crlf"}, \emph{"binary"}, or \emph{"auto"}. Default to \emph{"auto"}.}

\item{open}{Logical. Should connection be opened once established? Default to
TRUE}
}
\value{
A serial interface connection to the SDS011 device via USB
}
\description{
Connect to the SDS011 air particulate sensor connected via USB
}
\examples{
\dontrun{
  sds011_connect()
}

}
