% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds011.R
\name{sds011_read}
\alias{sds011_read}
\title{Read from a serial interface connection to the SDS011 device}
\usage{
sds011_read()
}
\value{
Readings from a serial interface connection to the SDS011 device
}
\description{
Read from a serial interface connection to the SDS011 device
}
\examples{
\dontrun{
  sds011_read()
}

}
