% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds011.R
\name{sds011_connect}
\alias{sds011_connect}
\title{Connect to the SDS011 air particulate sensor connected via USB}
\usage{
sds011_connect(open = TRUE)
}
\arguments{
\item{open}{Logical. Should connection be opened once established? Default to
TRUE}
}
\value{
A serial interface connection to the SDS011 device via USB
}
\description{
Connect to the SDS011 air particulate sensor connected via USB
}
\examples{
\dontrun{
  sds011_connect()
}

}
